Search page issue fixed

+++ WeatherAppTPNooro/Models/Location.swift
@@ -11,8 +11,8 @@ struct Location: Codable {
     var name: String
     var region: String
     var country: String
-    var lat: Double
-    var lon: Double
+    var lat: Float
+    var lon: Float
 }
 
 // MARK: - Equatable
